Removed unnecessary code duplication in State
While developing support for external types, I started looking into
this part of code in State that looked pretty bad.

It turns out it was all mostly duplicated code, as in exception of
one case any property type would have exactly the same code, as the bulk
of the struct <--> binary transformations are now done by
PropertyComponent.

The only slight difference is in L210, which throws an error in case
of the enum value is unrecognized. This may also be removed in the
future, or delegated to PropertyComponent.