Support size-less embedded strings
In AutoAPI L11 we changed the binary protocol for the strings embedded
in custom data types. Instead of storing the string size in an
additional mandatory `<string_name>_size` item, we now just prepend the
string with a 2-byte header with the string size.

For first-tier strings, that is properties of type string, we still
derive the string size from the property size header.