Fix a bug in AutoApi.Command.to_bin/3
The `to_bin/3` function that is autogenerated in L11 for each capability
command module behaves a bit differently than the ones that were
manually implemented in L10 such as it already includes the preamble
with the capability id.

The "old" `AutoApi.Command.to_bin/3` utility function used to add this
ID preamble before returning the binary command, therefore breaking the
binary protocol as the preamble got added twice.

This commit fixes that.

Also the `AutoApi.Command.capability_name` type got removed as it was
incomplete and should have been updated manually. Now it's just an alias
for the `atom` type. In the future we will autogenerate it somehow in
`AutoApi.Capability` module.