Allows shorthand syntax for Command.to_bin
While using it I noticed that wrapping the value in an
`PropertyComponent` struct is a bit cumbersome and also most of the time
not really necessary as the commands usually require only the value.

This commit updates `Command.to_bin` `:set` to accept also a Keyword list
of simple values rather than requiring always a PropertyComponent.